Add new FileUtilites class
Fix Undo for refresh btn click

Fix delete not working on barcodes
